Add missing 404 page
