+++ src/App.tsx
@@ -3,6 +3,7 @@ import NavBar from "@/components/NavBar";
 import ThemeProvider from "@/components/ThemeProvider";
 import { ErrorBoundary } from "react-error-boundary";
 import { Route, Routes, useLocation } from "react-router-dom";
+import ErrorComponent from "./components/ErrorComponent.tsx";
 import ScrollToTop from "./components/ScrollToTop.tsx";
 import AboutPage from "./pages/AboutPage";
 import CardPage from "./pages/CardPage";
@@ -31,7 +32,17 @@ function App() {
               <Route path="/sets" element={<SetPage />} />
               <Route path="/syntax" element={<SyntaxPage />} />
               <Route path="/random" element={<RandomPage />} />
-              <Route path="*" element={<div>This page is missing</div>} />
+              <Route
+                path="*"
+                element={
+                  <ErrorComponent
+                    error={{
+                      name: "404 Error",
+                      message: "This page does not exist.",
+                    }}
+                  />
+                }
+              />
             </Routes>
           </main>
           <Footer />
