+++ src/pages/FallBackPage.tsx
@@ -1,49 +1,20 @@
+import ErrorComponent from "@/components/ErrorComponent";
 import Footer from "@/components/Footer";
 import NavBar from "@/components/NavBar";
-import { Button } from "@/components/ui/button";
 import { useEffect } from "react";
-import { useErrorBoundary } from "react-error-boundary";
-import { TbFaceIdError } from "react-icons/tb";
 import { useNavigate } from "react-router-dom";
 
 export default function FallBackPage({ error }: { error: Error | null }) {
-  const { resetBoundary } = useErrorBoundary();
   const navigate = useNavigate();
 
   useEffect(() => {
-    if (!error) {
-      navigate("/");
-    }
+    if (!error) navigate("/");
   }, [error, navigate]);
-
-  const handleReset = () => {
-    resetBoundary();
-    navigate("/");
-  };
 
   return (
     <div className="flex min-h-screen flex-col">
       <NavBar />
-      <div
-        className="m-4 flex flex-1 flex-col items-center justify-center px-4 py-8"
-        role="alert"
-      >
-        <div className="max-w-lg rounded-lg p-6 text-center shadow-[0_0_15px_5px] shadow-shadow">
-          <TbFaceIdError
-            size={"7rem"}
-            className="mx-auto mb-4 text-muted-foreground"
-          />
-          <div data-test-id="error-message">
-            <h1 className="mb-2 text-xl font-semibold">
-              Something Went Wrong:
-            </h1>
-            <p className="mb-4">{error ? error.message : "Unknown error"}</p>
-          </div>
-          <Button onClick={handleReset} data-test-id="error-reset-button">
-            Go Back
-          </Button>
-        </div>
-      </div>
+      <ErrorComponent error={error} />
       <Footer />
     </div>
   );
